ft: main file refactors
