+++ backend/src/index.js
@@ -1,45 +1,21 @@
 import express from "express";
-import authRoutes from "./routes/auth.route.js";
-import messageRoutes from "./routes/message.route.js";
-import postRoutes from "./routes/post.route.js";
-
 import dotenv from "dotenv";
-import { connectDB } from "./lib/db.js";
 import cookieParser from "cookie-parser";
-import cors from "cors";
+
+import { connectDB } from "./lib/db.js";
 import { app, server } from "./lib/socket.js";
+import { corsOptions } from "./lib/corsOptions.js";
+import appRoutes from "./routes/index.js";
 
 dotenv.config();
 
 const PORT = process.env.PORT;
 
-const allowedOrigins = ["http://localhost:5173", "https://vuga.onrender.com"];
-
 app.use(express.json());
 app.use(cookieParser());
 
 // CORS configuration
-app.use(
-  cors({
-    origin: function (origin, callback) {
-      // Allow requests with no origin (like mobile apps or curl requests)
-      if (!origin || allowedOrigins.includes(origin)) {
-        callback(null, true);
-      } else {
-        callback(new Error("Not allowed by CORS"));
-      }
-    },
-    credentials: true,
-    methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
-    allowedHeaders: [
-      "Content-Type",
-      "Authorization",
-      "X-Requested-With",
-      "Accept",
-    ],
-    exposedHeaders: ["Set-Cookie"],
-  })
-);
+app.use(corsOptions);
 
 // Add security headers
 app.use((req, res, next) => {
@@ -47,9 +23,7 @@ app.use((req, res, next) => {
   next();
 });
 
-app.use("/api/auth", authRoutes);
-app.use("/api/messages", messageRoutes);
-app.use("/api/posts", postRoutes);
+app.use(appRoutes);
 
 server.listen(PORT, () => {
   console.log(`Server is listening on port ${PORT}`);
